Reset namespace filter on cancel
If the user changes inputs in the namespace filter dialog but closes
the dialog without submitting, reset the radios and checkboxes to
match the last submitted state (or the default state, if nothing has
been submitted).

Also, keep track of the current state of the filter via a prop, so
we can default back to it when needed.

Bug: T276261
Change-Id: I876ff9c17dc44f8c0db0cfe0bf5ad8dbbcf6be4a