Drop MediaInfoMediaSearchSupportedFilterParams config
IIRC, this was introduced to make PHP aware of which filters
were in use (back when they still lived in a json file that
only JS would use) in order to validate those params.

Another beneficial use of this, ot some point, was for
MediaSearch, where this would've allowed switching off
filters when that dependency is not available
(e.g. haslicense)

However, since the config have now moved into a code-
generated PHP file instead of the original JSON, and the
PHP is actively using it already, we no longer need this
to be configurable. PHP already knows which exist
(MediaSearchOptions::ALL_FILTERS)
Better yet, it already checks their availability per type
(via MediaSearchOptions::getOptions) so we've already
surpassed what this was originally introduced for.
The other benefit is also no longer needed, since the
config option are now dynamically generated and can (and
do) exclude options for which a dependency is not available.

One could argue that it *could* be convenient to allow
these to be configurable, but I think it only causes
confusion, because some simply don't apply for some types.
If there is a need for these to be configurable, we'll need
to make them more granular.
For now, I think we should get rid of them - they no longer
serve any purpose, so we can make the code just a little
bit simpler.

Change-Id: I9ae032dcd8e03e078c68420ca91c482e429e13b4