Don't show "invalid search" message when request is aborted by user
This problem had a previously-attempted fix at:
https://gerrit.wikimedia.org/r/c/mediawiki/extensions/WikibaseMediaInfo/+/673554.

However this didn't actually fix the problem in production due to a
discrepancy in the behavior of $.get (used in local development) and
mw.Api.prototype.get (used in production). The latter adds an error code
argument before the object containing the details we care about; this
meant that in production aborted requests were never properly distinguished
from real errors and the "invalid search" message was always shown.

Unfortunately, fixing the error handling in production will break it
in local development mode unless we can find a way to ensure consistency
between the request behaviors in both environments.

Bug: T277714
Change-Id: I06e959c0db25a48e183cc445b93a2157fee246db