Clear timeout for debounced lookup result request on search
The autocompleteLookupHandler debounces requests for lookup results
that are sent while the user types in the search input.

This change makes sure that the timeout is always cleared when a new
search is submitted. Otherwise, it is possible for the user to wind
up in a state where they have typed in a search term very quickly,
hit "enter", and then the autocomplete suggestions menu expands
and covers the results, and does not go away automatically.

Bug: T273665
Change-Id: I0b06536449b136d805705c75190fbf76467ebe29