Handle "missing mediainfo tabs" errors
According to our error logs, the initialization script for the MediaInfo
Filepage UI often fails because the .wbmi-tabs can't be found in the DOM.
This is almost certainly due to some unsupported gadget, user script, or
browser extension.

In order to avoid spamming the logs due to an error that seems to be
outside our control, this change simply wraps the infuseTabs() call in
a try / catch block. If infusion fails, the script logs the error to the
user's console and bails early.

Bug: T262470
Change-Id: I85fa9af91c7b2f4ff6b6c31e9d4eae3f056702d1