Re-enable "mark as prominent" in UW
This was causing the ItemWidget to always be disabled
(also not showing the "mark as prominent" buttons)
when we simply didn't want to publish controls (in the
case of UW)

The other case where we wouldn't want the controls is
when a page is protected, but that's already covered
in the bootstrapping, and will - in addition to setting
`showControls` to false - also set `disabled` false.
It is thus useless to also check for showControls in
this case, because the other condition will also be false.

Bug: T285579
Change-Id: I3b3e1d5f738cb5cd8fdb4c33d9a0cf364ee1e94c