Account for matches with 0 < score < 1 in normalization
Such matches could safely be ignored back when this workaround
was first concocted, because the scores were fairly large.
Nowadays, most boosts are in the 0.0001 to 0.1 range, and
many matches tend to remain below zero, and we effectively
end up normalizing those scores back down to 0.

This works around it differently in a way that scores below
zero also remain intact.

Bug: T294953
Change-Id: I7105c8f67cdea283ca1e690ee4a9d2b4a285d08e