Fix rendering
The existing code had a few naive omission that,
given convoluted circumstances, could result in
wrong or multiple nodes being matched between the
old & new rendered versions.

I'm sure this is still incomplete - the premise of
using a template instead of code makes it impossible
to accurately track individual nodes - but it's
less incomplete than what we used to have.

Bug: T296616
Change-Id: I6f90ccc1db3b251a4860a9cbaa45ea5c42a082f7