Refactor node matching
Instead of trying to find similar nodes (by multiple criteria)
in advance and then matching them up later on, I'm going to do
both combined. If not combined, we either have to choose between:
- very lenient criteria so we have many nodes to pick from, but
  hard to know which one really is the best (without repeating
  the same criteria)
- strict criteria (potentially dropping the actual result, in
  case there are multiple nodes that are very much alike)

By combining them, we can always be as strict as we want to be
in finding the best possible similar nodes.

In practice: the old code found that, between both old & new
DOM versions, both the qualifiers & references sections were
similar nodes for one another (they are essentially both
snaklists, after all.) But when moving from read mode to edit
mode, it was found that the old qualifiers div looked more
like the new references div than the old references diff did
(because it was missing one less node (the "add" button))
The old qualifiers div ended up correctly matching up with the
new qualifiers div, but the references div had no match and
it failed to properly merge both versions.

The way the new code handles this is smarter: it'll still
look for both similar & dissimilar nodes & cross-reference
matching nodes, but it'll only pick from what's still
available: by the time it comes to the new references div,
we will already have used up the old qualifiers div, and
find that the old references div is the best remaining match.

Bug: T297171
Change-Id: Ic94bc728785100a040f1502b2125f4c51df53c85