Update minimum_should_match for deprecations in 7
In elasticsearch 7 bool queries with should clauses that don’t need
to access the score will no longer set their minimum_should_match to
1. This effects any bool query that is referenced inside the filter of
another bool query. It's not always obvious when a query will be used
in a filter, make this more explicit by setting the minimum_should_match
based on the use case.

Bug: T288765
Change-Id: Ibcbc4de3476bcb2cdb961841abaf3a6da6e8ed36