Only manipulate mediainfo content
The "normal" rendering procedure creates a new title
for each slot, and then outputs its content.

For MediaInfo, we wanted to change that a bit:
- the SDC content should not be after wikitext content,
  but somewhere "around" the rest of the content
  (captions + wikitext in one tab; statements in another)
- we don't need that title

The current regex to find the slot title applies to any
slot. It works fine because there aren't any others (right
now), but once that changes, it would also remove titles
of other slots, which is obviously not desirable.

Change-Id: I083070624131bd7593c1ac0b3a2cf2495cdca9fa