Reduce logspam from "Revision X belongs to Y instead of expected Z"
That basic way of logging the exception created a lot of noise in
logstash because for every combination of revisionId and entityId it
would create a separate message that would have to be ignore separately.
This commit introduces a way of logging the error message that makes it
easy to consolidate and ignore in logstash, if needed.

It should probably also not be a warning, and probably be in a dedicated
"WikibaseMediaInfo" log channel, not the "Wikibase" one.

Bug: T193690
Depends-On: Id613b5177e133fdbaaf0e3cb2d228ebbabbca09a
Change-Id: I758fb5820a24986edc52963609de45dc9cf30da7