Make PHPUnit tests work in non-English environments
Some of the tests had hard-coded assumptions about the language of the
dev environment. Instead it's often better to let tests use the qqx
fake language code that returns message keys as-is without doing
anything expensive with the actual localization system. I18n is not
what these tests are about anyway. Just being able to find the correct
message key is enough.

I'm also using the opportunity to streamline some of the slightly
overcomplicated mock setup. These hard-coded maps of "expected" entity
ids appears to be pointless when there is nothing that would happen in
case an entity id is "unexpected", other than returning an empty string.
We can as well return any string then.

Change-Id: I04c67bbd38c760dbddf6ff7fee416d536ae3753b