Support making Wikibase dataProviders static
In I2b0873c7e54513f84011eaddc34d960b8490ceb1, the abstract function
`getEntity` was made static as part of the work to make all PHPUnit
dataProvider functions static.
`MediaInfoChangeOpDeserializerTest` implements this function, which
must match the declaration in the abstract classes.

So that the change in Wikibase could be merged, the static/non-static
mismatch was avoided by temporarily removing usage of the affected
abstract classes. This commit reverts that temporary change, and makes
`getEntity` static.

Bug: T337154
Depends-On: I2b0873c7e54513f84011eaddc34d960b8490ceb1
Change-Id: I2d54d97c5b48c39ea32fedf0b90361027fe36422