style: adjust shrink behavior of statements
This fixes an issue where at certain viewport sizes for Vector-2022, and
any viewport size in Timeless, the trash can icon could overflow for the
top-level statements. This fixes that by ensuring that the "mark as
prominent" text has a flex-shrink of 0, while the label has a
flex-shrink of 1, as we'd rather wrap the label if needed.

This additionally fixes an issue with the constraints violation icon
overlapping with the "mark as prominent" text, as it was positioned
absolutely instead of relatively, causing it to be ignored in the normal
document flow.

Bug: T404158
Bug: T404258
Change-Id: Ia79c39b6f0f3e441db5cc4fbca86229b973c62a4